JBPM-6920 (Stunner) - Align and distribute gets broken after an exception
https://issues.jboss.org/browse/JBPM-6920

Change in how the align and distribution indexation works. It used to
work by listening to changes in node attributes and updating it.
Most of the errors seemed to be caused due to missed updates due to a
change not been listened to.
To solve this, before a node drag event, all nodes and their values are
reindexed.